python: Set Session and UiProxy as globals for the embedded interpreter.
Setting Session and UiProxy as locals does not work!
When executing a script, the globals() inside a function are missing the
additional stuff like imported modules, function pointer and also everyhthing
provided by the locals dict. Setting Session and UiProxy in addition to
py::globals() as globals did the trick. See:
https://medium.com/just-me-me-programming-life/python-c-and-symbols-4628fb71a257