CMake: Use import targets for pkg-config libs.
It is easyer to pass all libs and search paths to the CheckSigrokFeatures
script with imported targets. Also we don't have to check, if we need the
static libs.
This should also fix the build issues on Fedora 32 from issue #32.

Update function name in the sigrok check to lower case.