BaseChannel: Unfuse check when adding a signal to a fixed channel.
The check when adding signals with other mq/mq_flags than specified in e.g.
SourceSinkDevice to a fixed channel was defused in BaseChannel::add_signal()
to prevent a subsequent segfault.
This happens when sigrok drivers send unexpected mq/mq_flags (can be fixed in
the sigrok driver), but also for AC power supplies and loads which is a design
issue in SmuView right now that requires an API enhancement in libsigrok!
With this fix, AC psu/loads will have two signals (DC and AC).

Also add a Python test for fixed channels.

Fixes issue #35.