Change the meaning and handling of digits() from sigrok::Analog
The actual meaning of digits() is:
This property holds the number of significant digits after the decimal point,
if positive. When negative, exponent with reversed polarity that is necessary
to express the value with all digits without a decimal point.

This commit now also sets sensible default values for sr_digits, total_digits
and decimal_places for channles, signals and properties.

This commit adds tests to verify the new bahaviour.